Add header and cookie after request to control cache
We never want to cache a responses if a user is authenticated. We do
this my setting the Cache-Control header to No-Cache. The check of the
login_manager attribute on the current app is important so this code can
be shared across front and backend apps.

We also want to set a cookie so we can let logged in users bypass the
cache up front. The dm_loggged_in cookie will either be TRUE, FALSE or
not present. If it's not present (a user hits the app for the first time
for example), then we set it to FALSE here. Once they log in it gets set
to TRUE, and FALSE when they log out. See the user frontend.